feat(website): emote ticket dialog mock
